feat(usdtieredsto): tiered sto allows set any stablecoin we want
the module is now available to be pegged with other fiat currency than usd

BREAKING CHANGE: Contract is now available to be pegged with other fiat currency. Configure function
is now accepting the contract address of the stable coins, the addresses of the custom oracles and
the symbol of the denominated currency.